+++ src/east/__main__.py
@@ -2,7 +2,7 @@ import click
 import rich_click
 
 from .east_context import EastContext, east_group_settings
-from .system_commands import sys_setup
+from .system_commands import sys_setup, util
 from .workspace_commands import build, bypass, clean, flash, update
 
 rich_click.rich_click.MAX_WIDTH = 80
@@ -17,10 +17,11 @@ rich_click.rich_click.COMMAND_GROUPS = {
         },
         {
             "name": "System commands",
-            "commands": ["sys-setup"],
+            "commands": ["sys-setup", "util"],
         },
     ],
     "east update": [{"name": "Subcommands", "commands": ["west", "env", "toolchain"]}],
+    "east util": [{"name": "Subcommands", "commands": ["connect", "rtt"]}],
 }
 
 
@@ -65,6 +66,7 @@ cli.add_command(flash)
 cli.add_command(clean)
 cli.add_command(update)
 cli.add_command(sys_setup)
+cli.add_command(util)
 
 
 def main():