+++ src/east/system_commands/__init__.py
@@ -1,1 +1,2 @@
 from .sys_setup import sys_setup
+from .util_commands import util